fix: move short links to its own hook, mutation and api file
