+++ src/hooks/item.ts
@@ -39,8 +39,6 @@ import {
   buildItemParentsKey,
   buildItemThumbnailKey,
   buildItemsKey,
-  buildShortLinkKey,
-  buildShortLinksItemKey,
 } from '../config/keys';
 import { getOwnItemsRoutine } from '../routines';
 import { QueryClientConfig } from '../types';
@@ -456,26 +454,5 @@ export default (
       }),
 
     useItemFeedbackUpdates: itemWsHooks?.useItemFeedbackUpdates,
-
-    useShortLink: (alias: string | undefined) =>
-      useQuery({
-        queryKey: buildShortLinkKey(alias),
-        queryFn: () => {
-          if (!alias) {
-            throw new UndefinedArgument();
-          }
-          return Api.getShortLink(alias, queryConfig);
-        },
-        enabled: Boolean(alias),
-        ...defaultQueryOptions,
-      }),
-
-    useShortLinksItem: (itemId: string) =>
-      useQuery({
-        queryKey: buildShortLinksItemKey(itemId),
-        queryFn: () => Api.getShortLinksItem(itemId, queryConfig),
-        enabled: true,
-        ...defaultQueryOptions,
-      }),
   };
 };
