+++ src/mutations/index.ts
@@ -17,6 +17,7 @@ import memberMutations from './member';
 import itemMembershipMutations from './membership';
 import mentionMutations from './mention';
 import subscriptionMutations from './plan';
+import shortLinksMutation from './shortLink';
 
 const configureMutations = (queryConfig: QueryClientConfig) => ({
   ...itemMutations(queryConfig),
@@ -37,6 +38,7 @@ const configureMutations = (queryConfig: QueryClientConfig) => ({
   ...authenticationMutations(queryConfig),
   ...subscriptionMutations(queryConfig),
   ...itemPublishMutations(queryConfig),
+  ...shortLinksMutation(queryConfig),
 });
 
 export default configureMutations;
