+++ src/mutations/item.ts
@@ -1,11 +1,9 @@
 import {
-  AnyOfExcept,
   DiscriminatedItem,
   FolderItemExtra,
   ItemType,
   MAX_TARGETS_FOR_MODIFY_REQUEST,
   RecycledItemData,
-  ShortLinkPayload,
   ThumbnailSize,
   UUID,
 } from '@graasp/sdk';
@@ -24,21 +22,17 @@ import {
   buildItemChildrenKey,
   buildItemKey,
   buildItemThumbnailKey,
-  buildShortLinksItemKey,
   getKeyForParentId,
 } from '../config/keys';
 import {
   copyItemsRoutine,
   createEtherpadRoutine,
   createItemRoutine,
-  createShortLinkRoutine,
   deleteItemsRoutine,
-  deleteShortLinkRoutine,
   editItemRoutine,
   importH5PRoutine,
   importZipRoutine,
   moveItemsRoutine,
-  patchShortLinkRoutine,
   recycleItemsRoutine,
   restoreItemsRoutine,
   uploadFileRoutine,
@@ -120,82 +114,6 @@ export default (queryConfig: QueryClientConfig) => {
         onSettled: (_data, _error, { parentId }) => {
           const key = getKeyForParentId(parentId);
           queryClient.invalidateQueries(key);
-        },
-      },
-    );
-  };
-
-  // TODO: update success messages !
-  const usePostShortLink = () => {
-    const queryClient = useQueryClient();
-    return useMutation(
-      async (shortLink: ShortLinkPayload) =>
-        Api.postShortLink(shortLink, queryConfig),
-      {
-        onSuccess: (_data, variables) => {
-          notifier?.({
-            type: createShortLinkRoutine.SUCCESS,
-            payload: { message: SUCCESS_MESSAGES.CREATE_SHORT_LINK },
-          });
-          queryClient.invalidateQueries(
-            buildShortLinksItemKey(variables.item.id),
-          );
-        },
-        onError: (error: Error) => {
-          notifier?.({
-            type: createShortLinkRoutine.FAILURE,
-            payload: { error },
-          });
-        },
-      },
-    );
-  };
-
-  const usePatchShortLink = () => {
-    const queryClient = useQueryClient();
-    return useMutation(
-      async ({
-        alias,
-        shortLink,
-      }: {
-        alias: string;
-        shortLink: AnyOfExcept<ShortLinkPayload, 'item'>;
-      }) => Api.patchShortLink(alias, shortLink, queryConfig),
-      {
-        onSuccess: (data) => {
-          notifier?.({
-            type: patchShortLinkRoutine.SUCCESS,
-            payload: { message: SUCCESS_MESSAGES.EDIT_SHORT_LINK },
-          });
-          queryClient.invalidateQueries(buildShortLinksItemKey(data.item.id));
-        },
-        onError: (error: Error) => {
-          notifier?.({
-            type: patchShortLinkRoutine.FAILURE,
-            payload: { error },
-          });
-        },
-      },
-    );
-  };
-
-  const useDeleteShortLink = () => {
-    const queryClient = useQueryClient();
-    return useMutation(
-      async (alias: string) => Api.deleteShortLink(alias, queryConfig),
-      {
-        onSuccess: (data) => {
-          notifier?.({
-            type: deleteShortLinkRoutine.SUCCESS,
-            payload: { message: SUCCESS_MESSAGES.DELETE_SHORT_LINK },
-          });
-          queryClient.invalidateQueries(buildShortLinksItemKey(data.item.id));
-        },
-        onError: (error: Error) => {
-          notifier?.({
-            type: deleteShortLinkRoutine.FAILURE,
-            payload: { error },
-          });
         },
       },
     );
@@ -665,13 +583,10 @@ export default (queryConfig: QueryClientConfig) => {
 
   return {
     usePostItem,
-    usePostShortLink,
-    usePatchShortLink,
     usePostEtherpad,
     useEditItem,
     useRecycleItems,
     useDeleteItems,
-    useDeleteShortLink,
     useCopyItems,
     useUploadFiles,
     useUploadItemThumbnail,
